object_detection/yolov4/requirements.txt
opencv-python==4.1.1.26
lxml
tqdm
tensorflow #==2.1.0
absl-py
easydict
matplotlib
pillow
tensorflow_addons==0.9.1
pandas
paramiko
openpyxl